feat: Add support for custom relay files
Previously, the only way to specify relays for import and blastr was to list them directly in the .env file. This commit introduces the ability to specify a file path for the relays instead. Two example JSON files have been added to the .gitignore, and instructions for using them have been added to the README. This feature provides a more flexible and scalable way to manage relays.